requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
autopep8==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
fastapi==0.115.0
frozenlist==1.5.0
greenlet==3.2.3
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.0.3
multidict==6.2.0
packaging==24.2
propcache==0.3.0
psycopg2==2.9.10
pycodestyle==2.14.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==10.4
yarl==1.18.3
